+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [UNRELEASED] neptune-optuna 0.10.0
+
+### Features
+- Add support for multi-objective training ([#14](https://github.com/neptune-ai/neptune-optuna/pull/14))
+
 ## neptune-optuna 0.9.15
 
 ### Changes
@@ -6,13 +11,16 @@
 ## neptune-optuna 0.9.14
 
 ### Features
+
 - Mechanism to prevent using legacy Experiments in new-API integrations ([#6](https://github.com/neptune-ai/neptune-optuna/pull/6))
 
 ### Fixes
+
 - Added more checks to prevent exceptions ([#11](https://github.com/neptune-ai/neptune-optuna/pull/11))
-- Disabled ploting parameters importance on failure ([#10](https://github.com/neptune-ai/neptune-optuna/pull/10))
+- Disabled plotting parameters importance on failure ([#10](https://github.com/neptune-ai/neptune-optuna/pull/10))
 
 ## neptune-optuna 0.9.13
 
 ### Fixes
+
 - Handling exceptions and trial pruning ([#5](https://github.com/neptune-ai/neptune-optuna/pull/5))